test(runner): fold stream marker regression into runner suite
